net: stmmac: add TX HW Time-stamping error fallback mode
It is possible that the returned T/S is not valid due to descriptor error.
So, we add check for detecting descriptor error in write-back Tx
descriptor.

If HW TX TS is not valid, it is still better to get the current PTP clock
to approximate the TX frame TS instead of not updating the TS which may
be used by caller as stale and old TS from the past descriptor and cause
issue to application.

Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>